+++ RUFAS/routines/field/manager/tillage_schedule.py
@@ -3,6 +3,7 @@ from typing import List
 from RUFAS.routines.EEE.enums import TillageImplement
 from RUFAS.routines.field.manager.events import TillageEvent
 from RUFAS.routines.field.manager.schedule import Schedule
+from RUFAS.util import Utility
 
 
 class TillageSchedule(Schedule):
@@ -60,16 +61,16 @@ class TillageSchedule(Schedule):
     ):
         super().__init__(name, years, days, pattern_skip, pattern_repeat)
 
-        self.tillage_depths = self._elongate_list(tillage_depths, len(years))
-        self.incorporation_fractions = self._elongate_list(incorporation_fractions, len(years))
-        self.mixing_fractions = self._elongate_list(mixing_fractions, len(years))
+        self.tillage_depths = Utility.elongate_list(tillage_depths, len(years))
+        self.incorporation_fractions = Utility.elongate_list(incorporation_fractions, len(years))
+        self.mixing_fractions = Utility.elongate_list(mixing_fractions, len(years))
 
         self.implements = [TillageImplement(implement) for implement in implements]
-        self.implements = self._elongate_list(self.implements, len(years))
+        self.implements = Utility.elongate_list(self.implements, len(years))
 
         self._validate_tillage_parameters()
 
-        self._validate_pattern_parameters()
+        Utility.validate_pattern_parameters(self.name, self.pattern_skip, self.pattern_repeat)
 
     def _validate_tillage_parameters(self) -> None:
         """
@@ -146,7 +147,7 @@ class TillageSchedule(Schedule):
             List of all tillage events that will happen over the course of this TillageSchedule.
 
         """
-        all_tilling_years = self._repeat_pattern(self.years, self.pattern_skip, self.pattern_repeat)
+        all_tilling_years = Utility.repeat_pattern(self.years, self.pattern_skip, self.pattern_repeat)
         all_tilling_days = self.days * (self.pattern_repeat + 1)
         all_tillage_depths = self.tillage_depths * (self.pattern_repeat + 1)
         all_incorporation_fractions = self.incorporation_fractions * (self.pattern_repeat + 1)