+++ RUFAS/util.py
@@ -3,6 +3,7 @@ import enum
 import os
 import re
 import shutil
+from copy import copy
 from pathlib import Path
 from typing import Any, Callable, Dict, List, Optional, Tuple
 
@@ -644,3 +645,94 @@ class Utility:
         result_df.to_csv(output_csv_path, index=False)
 
         shutil.rmtree(saved_csv_working_folder)
+
+    @staticmethod
+    def elongate_list(list_to_elongate: List[Any], reference_list_length: int) -> List[Any]:
+        """
+        Takes a list and lengthens it to match the length of the reference list, if the original length was 1.
+
+        Parameters
+        ----------
+        list_to_elongate : List[Any]
+            List to be extended if its length is 1.
+        reference_list_length : int
+            Length of that the list should be extended to, if it its original length is 1.
+
+        Returns
+        -------
+        List[Any]
+            The elongated list.
+
+        Notes
+        -----
+        In the context of Schedule-descendant classes, the reference list length will always be the length of the years
+        list.
+
+        """
+        if len(list_to_elongate) != 1:
+            return list_to_elongate
+        elongated_list = list_to_elongate * reference_list_length
+        return elongated_list
+
+    @staticmethod
+    def repeat_pattern(pattern: List[int], skip: int = 0, repeat: int = 0) -> List[int]:
+        """
+        Takes a pattern of numbers and repeats the pattern of differences between the numbers for a specified number of
+        repetitions, skipping over specified gaps between repetitions.
+
+        Parameters
+        ----------
+        pattern : List[int]
+            The pattern to be repeated.
+        skip : int
+            Number of steps to skip between repeats (0 if no steps should be skipped).
+        repeat : int
+            Number of times pattern should be repeated.
+
+        Returns
+        -------
+        List[int]
+            The full repeated pattern of numbers.
+
+        Examples
+        --------
+        >>> repeat_pattern([1, 3, 5], 1, 2)
+        [1, 3, 5, 7, 9, 11, 13, 15, 17]
+
+        >>> repeat_pattern([1, 3, 5], 0, 1)
+        [1, 3, 5, 6, 8, 10]
+
+        >>> repeat_pattern([2, 3, 7], 3, 2)
+        [2, 3, 7, 11, 12, 16, 20, 21, 24]
+
+        """
+        differences = [skip + 1]
+        in_pattern_differences = range(1, len(pattern[1:]) + 1)
+        for difference in in_pattern_differences:
+            differences.append(pattern[difference] - pattern[difference - 1])
+
+        full_pattern = copy(pattern)
+        differences_index = 0
+        number_of_new_values = range(repeat * len(pattern))
+        for _new_value in number_of_new_values:
+            full_pattern.append(full_pattern[-1] + differences[differences_index])
+            differences_index += 1
+            differences_index %= len(pattern)
+        return full_pattern
+
+    @staticmethod
+    def validate_pattern_parameters(name: str, pattern_skip: int, pattern_repeat: int) -> None:
+        """
+        Checks the pattern skip and repeat parameters, if they are not correct raises errors.
+
+        Raises
+        ------
+        ValueError
+            If the skip is < 0.
+            If the repeat is < 0.
+
+        """
+        if pattern_skip < 0:
+            raise ValueError(f"'{name}': expected pattern skip to be >= 0, received '{pattern_skip}'.")
+        if pattern_repeat < 0:
+            raise ValueError(f"'{name}': expected pattern repeat to be >= 0, received '{pattern_repeat}'.")