+++ tests/test_util.py
@@ -862,3 +862,119 @@ def test_combine(
     mock_list_dir.assert_called_once_with(saved_csv_working_folder)
 
     mock_rmtree.assert_called_once_with(saved_csv_working_folder)
+
+
+@pytest.mark.parametrize(
+    "test_list,length,expected",
+    [
+        ([], 3, []),
+        ([], 0, []),
+        ([1, 2], 1, [1, 2]),
+        ([1.0, 2.0], 5, [1.0, 2.0]),
+        (["test"], 4, ["test", "test", "test", "test"]),
+        ([3], 1, [3]),
+        ([5], 5, [5, 5, 5, 5, 5]),
+    ],
+)
+def test_elongate_list(test_list: List[Any], length: int, expected: List[Any]) -> None:
+    """Check that lists are elongated correctly."""
+    actual = Utility.elongate_list(test_list, length)
+    assert actual == expected
+
+
+@pytest.mark.parametrize(
+    "pattern, skip, repeat, expected",
+    [
+        (
+            [1, 3, 5],
+            1,
+            3,
+            [
+                1,
+                3,
+                5,
+                7,
+                9,
+                11,
+                13,
+                15,
+                17,
+                19,
+                21,
+                23,
+            ],
+        ),
+        ([1, 3, 5], 0, 1, [1, 3, 5, 6, 8, 10]),
+        (
+            [2, 3, 7],
+            3,
+            2,
+            [
+                2,
+                3,
+                7,
+                11,
+                12,
+                16,
+                20,
+                21,
+                25,
+            ],
+        ),
+        ([2, 3, 7], 0, 0, [2, 3, 7]),
+        ([2], 0, 0, [2]),
+        ([2], 3, 1, [2, 6]),
+        ([2], 0, 5, [2, 3, 4, 5, 6, 7]),
+        (
+            [2, 3, 3],
+            2,
+            3,
+            [
+                2,
+                3,
+                3,
+                6,
+                7,
+                7,
+                10,
+                11,
+                11,
+                14,
+                15,
+                15,
+            ],
+        ),
+        ([1, 1], 0, 4, [1, 1, 2, 2, 3, 3, 4, 4, 5, 5]),
+        ([1, 1, 3], 3, 1, [1, 1, 3, 7, 7, 9]),
+        ([], 0, 0, []),
+        ([], 3, 7, []),
+    ],
+)
+def test_repeat_pattern(pattern: List[int], skip: int, repeat: int, expected: list) -> None:
+    """Tests that repeat_pattern correctly repeats patterns."""
+    assert Utility.repeat_pattern(pattern, skip, repeat) == expected
+
+
+@pytest.mark.parametrize(
+    "name,skip,repeat,expected,error",
+    [
+        ("test_1", -1, 1, "'test_1': expected pattern skip to be >= 0, received '-1'.", True),
+        (
+            "test_2",
+            1,
+            -1,
+            "'test_2': expected pattern repeat to be >= 0, received '-1'.",
+            True
+        ),
+        ("test_3", 1, 1, "Nothing", False)
+    ],
+)
+def test_validate_pattern_parameters(name: str, skip: int, repeat: int, expected: str, error: bool) -> None:
+    """Tests that errors are correctly raised by Schedule when invalid"""
+    if error:
+        with pytest.raises(ValueError) as e:
+            Utility.validate_pattern_parameters(name, skip, repeat)
+        assert str(e.value) == expected
+    else:
+        Utility.validate_pattern_parameters(name, skip, repeat)
+        assert True